requirements.txt
discord.py==2.3.2
grequests==0.7.0
PyNaCl==1.5.0
python-dotenv==1.0.1
prettytable==3.10.0
